Fix exception thrown when !relocate was used
`getNewLocationForSheep` replaces a block in the world to give players
a platform, but only when there's water or other 'hazard'. It seems like
replacing blocks like that needs to be done in the main thread, and our
methods are run asynchronously by default.

Calling it inside `Bukkit.getScheduler().runTask` instead fixes it.