Implement simpler parameters for !tnt command
Added the option to specified predefined values for yaw, pitch, ttl and
power/distance. The values come from words that are easier to understand
for players (e.g. "north" for yaw, "up" for pitch, etc.).

Shorthands for the words are also supported (e.g. "q" for "quick").

Default values are provided for pitch, distance and ttl. Yaw is always
required.

Finally, the words do not need to be in a given sequence. Since all of
them are different, we can parse them in any order.